bugfix/fixing-error-middlewares
