+++ app.js
@@ -4,24 +4,17 @@ const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
 const { errors } = require('celebrate');
 const helmet = require('helmet');
-const { rateLimit } = require('express-rate-limit');
 
 const routes = require('./routes/routes');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 const error = require('./middlewares/error');
 const cors = require('./middlewares/cors');
+const limiter = require('./middlewares/limiter');
 
 const { NODE_ENV, DB_ADDRESS } = process.env;
 
 const app = express();
 const PORT = 3000;
-
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // 15 minutes
-  max: 100, // Limit each IP to 100 requests per `window` (here, per 15 minutes)
-  standardHeaders: true, // Return rate limit info in the `RateLimit-*` headers
-  legacyHeaders: false, // Disable the `X-RateLimit-*` headers
-});
 
 app.use(helmet());
 app.use(cors);
